api/requirements.txt
asyncpg==0.30.0
boto3==1.38.13
fastapi==0.115.12
pydantic==2.11.4
pydantic_settings==2.9.1
python-dotenv==1.1.0
python_jose==3.4.0
Requests==2.32.3
starlette==0.46.2
